// style-vision-app/frontend/src/types/ClothingItem.ts
export interface ClothingItem {
  id: number;
  image_path: string;
  description: string;
  garment_type: string;
  style: string;
  material: string;
  color_palette: string;
  pattern: string;
  season: string;
  occasion: string;
  consumer_profile: string;
  trend_notes: string;
  continent: string;
  country: string;
  city: string;
}